pyproject.toml
[build-system]
requires = ["scikit-build-core>=0.3.3"]
build-backend = "scikit_build_core.build"

[project]
name = "django-collectd-rest"
version = "0.2.4"
authors = [
  {name = "Matwey V. Kornilov", email = "matwey.kornilov@gmail.com"},
]
description = "A simple Django application to demonstrate RRD plots generated by collectd or any other rrd data"
readme = "README.md"
license = {text = "BSD-2-Clause"}
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Environment :: Web Environment",
  "Framework :: Django",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: BSD License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python",
  "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
  "Topic :: Internet :: WWW/HTTP",
]
requires-python = ">= 3.7"
dependencies = [
  "djangorestframework",
]

[project.optional-dependencies]
dev = [
  "mock",
  "pytest",
  "pytest-django",
]

[project.urls]
Repository = "https://github.com/matwey/django-collectd-rest.git"
Issues = "https://github.com/matwey/django-collectd-rest/issues"

[tool.scikit-build]
cmake.verbose = true
wheel.packages = ["src/collectd_rest"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
django_find_project = false
pythonpath = ". src"
testpaths = [
    "src/collectd_rest",
    "tests",
]
